third_party/WebKit/LayoutTests/fast/forms/calendar-picker/calendar-picker-datetimelocal.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type=datetime-local id=datetimelocal value="2011-12-31T12:53:45"> 
<script>
description("Tests if calendar picker works for datetime-local as expected.");

debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

openPicker(document.getElementById('datetimelocal'), test);

function test() {
    shouldBeEqualToString('selectedValue()', '2011-12-31');
    shouldBeEqualToString('highlightedValue()', '2011-12-31');

    eventSender.keyDown('ArrowRight');
    shouldBeEqualToString('highlightedValue()', '2012-01-01');

    removeCommitDelay();
    eventSender.keyDown('Enter');
    shouldBeEqualToString('document.getElementById("datetimelocal").value', '2012-01-01T12:53:45');

    finishJSTest();
}

</script>
</body>
</html>
